+++ CHANGELOG.md
@@ -39,6 +39,7 @@
 - PIM-10377: Change Elastic Search field Limit for everyone
 - PIM-10384: Fix cannot build EE standard due to "can't resolve '@akeneo-pim-enterprise/onboarder'" error
 - PIM-10417: Increase ImageMagick disk resource
+- PIM-10408: Fix bad product version when importing a product table value with wrong option case
 
 ## Improvements